tls: Make SanMatcher::match consistent (#39513)
Function was implemented with a mix of `const
GENERAL_NAME*` and `GENERAL_NAME const*`. I expect the latter was the
intention, a pointer to an immutable struct, not an immutable pointer to
a mutable struct. (Arguably the intention is *both* kinds of const, but
there is no meaningful value in the pointer being immutable.)
The current implementation also leads to many warnings in the gcc build
of the form
```
./source/common/tls/cert_validator/san_matcher.h:27:16: warning: ‘virtual bool Envoy::Ssl::SanMatcher::match(const GENERAL_NAME*, const Envoy::StreamInfo::StreamInfo&) const’ was hidden [-Woverloaded-virtual]
    27 |   virtual bool match(GENERAL_NAME const* general_name, const StreamInfo::StreamInfo&) const {
```

Signed-off-by: Raven Black <ravenblack@dropbox.com>