Use static linking for fuzz coverage CI targets
Static linking helps to work around the issue with LLVM/Clang
source-based coverage (see llvm/llvm-project#32849). On the flip side,
the coverage build and test run will take a bit longer with this change.

We already did this change for the regular test coverage, but we delayed
the switch for the fuzz coverage because fuzz coverage hit RBE backend
resource constraints and was OOMing during linking.

Since then we did a few things to mitigate the issue with resources:

1. We switched coverage runs from Google RBE backend to EngFlow which
   allows for somewhat larger workers
2. We started building fuzz targets without WASM - there are not fuzz
   targets that actually need a full WASM runtime and elimnating it
   speeds things up and reduces the footprint noticably.

With the above two changes I think we are ready to completely switch
to static linking for all coverage tests. It also has a benefit of
better cache re-use since we will be linking both coverage and fuzz
coverage targets statically.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>