ext_proc: Fix filter protocol configuration encoding with per-route config (#39650)
If processing mode exists in per-router configuration, it overrides
what's in the HTTP filter configuration.
The final processing mode configurations are stored in the filter
decoding_state_ and encoding_state_. Encoding them when sending those
modes to the ext_proc server.

This PR is a followup of https://github.com/envoyproxy/envoy/pull/38755,
which missed the per-route override part.

---------

Signed-off-by: Yanjun Xiang <yanjunxiang@google.com>