Adds note in the docs for ORCA Header Propagation (#39599)
Commit Message:
As discussed in #39382 this is something operators should be aware of
when using the ORCA LB policy.

So I added a note in the proto file for the ORCA LB policy to show up in
the docs.

Additional Description: N/A
Risk Level: N/A
Testing: N/A
Docs Changes: only docs
Release Notes: N/A
Platform Specific Features: N/A

---------

Signed-off-by: sotiris <sotiris.nanopoulos@reddit.com>