Explain why not using inet_pton for IPv6 addresses (#39686)
Commit Message: Explain why not using inet_pton for IPv6 addresses
Additional Description: It took me a while to figure out why we're using
the much more complicated and less performant `getaddrinfo()` for
`parseV6Address`, rather than `inet_pton`, which google searches
generally recommend using. Even the discussion history on the
responsible commits didn't really clarify, and `scope` as an explanation
was insufficient because the word is way too overloaded. This PR adds a
couple of lines of comments that make the reason for this choice clear
for the next confused person.
Risk Level: None, change is comment-only.
Testing: n/a
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

Signed-off-by: Raven Black <ravenblack@dropbox.com>