mutation_rules: replace proto data member with only required flags (#39764)
The mutation-rules Checker class keeps the entire Rules proto as a data
member, but only a few boolean flags may be accessed later. This PR
removes the data-member and introduces the flags as data members.
This will reduce any memory needed for the the regex fields in that
proto.

Risk Level: low
Testing: N/A
Docs Changes: N/A
Release Notes: N/A
Platform Specific Features: N/A

Signed-off-by: Adi Suissa-Peleg <adip@google.com>