Don't build API with libstdc++ by default
It seems that in the past there was a reason for that because we linked
some pre-built artifacts (api booster) that were built using libstdc++,
but it does not seem like we do that anymore.

As we are moving to managing toolchains and standard library via Bazel
it would be useful to limit the number of combinations of toolchain and
standard library we need to support to simplify things.

If we could eliminate this, we can stick to just two combinations:

1. gcc + libstdc++
2. llvm + libc++

libstdc++ is native to gcc and libc++ is native to llvm, so these
combinations make sense. And if we only need to support those two it
simplifies things a bit because we can get away and bundle toolchain
together with the standard library.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>