redis_proxy: Add support for per-upstream auth credentials
This adds the ability to use upstream redis clusters which each have different usernames and/or passwords. The credentials are keyed with an `address`, which must be a socket address and matches against the upstream based on the port along with the hostname or IP address.

Tested: Unit and integration tests, as well as manual testing and using in production at my company for a couple months

Fixes: https://github.com/envoyproxy/envoy/issues/23962

Signed-off-by: Ben Breslauer <bb8gh@github.com>