implement forwarding configuration in API Key Auth filter (#39763)
**Commit Message:** api: Implement forwarding configuration in API Key
Auth filter

**Additional Description:**
This feature provides control over what information is propagated to
upstream services. It allows forwarding the authenticated client
identity using a custom header, and also offers the option to remove the
API key from the request before forwarding. Both behaviors are optional
and can be configured separately.

**Risk Level:** Low

**Testing:** added several new unit tests for the feature in question as
well as a unit test for the new utility method, which removes a cookie
value from the request headers. Updated the configuration used in the
API Key Auth integration test accordingly.

**Docs Changes:** updated documentation for the API Key Auth Filter to
account for the new configuration fields introduced by this feature,
both in docs/root and in the respective API proto.

**Release Notes:** added to api_key_auth area
**Platform Specific Features:** n/a
**API Considerations:** added an optional field to the API Key Auth
filter configuration.

Closes #38690

---------

Signed-off-by: Leonor Figueira <leonor.figueira@tecnico.ulisboa.pt>
Co-authored-by: Raquel Rodrigues <raquel.c.rodrigues@tecnico.ulisboa.pt>