Make timeouts work in IntegrationStreamDecoder (#39801)
Commit Message: Make timeouts work in IntegrationStreamDecoder
Additional Description: Without this change, when an integration test is
using simulated time, IntegrationStreamDecoder can block indefinitely
despite there being a timeout, resulting in, if a thing is flaky in CI,
15 minute timeouts with no useful output, like
```
[ RUN      ] WhateverTest.ValidateWhatever
--- Test timed out ---
```

After this change, the failure is more useful, like
```
[ RUN      ] WhateverTest.ValidateWhatever
test/extensions/filters/http/some_filter/some_filter_integration_test.cc:184: Failure
Value of: response_decoder->waitForEndStream()
  Actual: false (Timed out (10000ms) waiting for end stream
IntegrationStreamDecoder state:
  saw_end_stream_=0
  saw_reset_=0
  headers_=':status', '400'
'date', 'Tue, 10 Jun 2025 19:10:58 GMT'
'server', 'envoy'

  continue_headers_=null
  body_=
  trailers_=null)
Expected: true
[ FAILED  ] WhateverTest.ValidateWhatever
```

This was modeled on how `FakeUpstream` does the same thing, with some
modifications to accommodate keeping the API unchanged.

This also revealed that the previous implementation of `waitForReset`
did not in fact validate that a reset was received (and was misused in
several places where a reset was not received). For these cases I have
added a `waitForAnyTermination`, since the tests in question don't
particularly care whether it's an end-stream or a reset.

I also made it fail before the timeout if the thing being waited for
couldn't happen (e.g. if a reset or end stream already happened then the
other one of those, or headers, or additional body, cannot happen.)

I also added more output about the state of the fake decoder on error,
so you can see how much of the output it *did* receive before the
timeout or unexpected termination, in the event of failure.

Risk Level: Test-only
Testing: Yes it is
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>