router: refactor the cluster validation step (#39874)
Commit Message: router: refactor the cluster validation step
Additional Description:
Prior to this PR, for every route the cluster-manager will create a
temporary map of the known clusters, and that will be used in the
validation step. This will incur a temporary memory and time complexity
of O(|clusters|).
This change removes the need for the temporary object, by directly
querying the cluster-manager if the cluster name is known or not.

Risk Level: low
Testing: N/A
Docs Changes: N/A
Release Notes: N/A
Platform Specific Features: N/A

---------

Signed-off-by: Adi Suissa-Peleg <adip@google.com>