init_manager: add option for updating watcher (#39914)
Adding an option to update the watcher that the
init manager notifies when it is done initializing. This will be used
for #39141 in next PR. This is needed for cases where the original
watcher is being discarded but the init targets are still initializing
and are part of an initialization process. Specifically for FCDS - this
is required in case an FCDS update is received while a previous listener
is warming. In that case, we don't want to allow the new listener to
warm up until all targets of the previous listener have also finished
warming up, because the FCDS updates are incremental
Risk Level: low
Testing: unit test
Docs Changes: none
Release Notes: none
Platform Specific Features: none

Signed-off-by: Ohad Vano <ohadvano@gmail.com>