xds: initial support for multiple xDS authorities config sources (with more tests) (#39781)
This is the same as #39556 (that was reverted in #39778 due to coverage
break) but with more tests.

Plumbing in the configuration parsing for the authorities-based
config-sources in the bootstrap file

Risk Level: low - unused path at the moment
Testing: Added unit-tests
Docs Changes: N/A
Release Notes: N/A
Platform Specific Features: N/A

Part of #24773.

---------

Signed-off-by: Adi Suissa-Peleg <adip@google.com>