lua: adds dynamicTypedMetadata() on streamInfo() (#39820)
## Description

This PR adds support for accessing `typed_filter_metadata` through the
stream info API. This allows Lua scripts to safely access
protobuf-encoded metadata, with proper unpacking of **Any** messages
into Lua tables. The implementation just extends the support for
`dynamicTypedMetadata()` which we added for `connectionStreamInfo()` in
the earlier PR.

---

**Commit Message:** lua: adds `dynamicTypedMetadata()` on `streamInfo()`
**Additional Description:** Added a new ``dynamicTypedMetadata()`` on
``streamInfo()`` which could be used to access the typed metadata from
HTTP filters, such as the ExtProc, etc.
**Risk Level:** Low
**Testing:** Unit tests and integration tests added
**Docs Changes:** Added
**Release Notes:** Added

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>