mobile: simplify fetch example (#39733)
**Commit Message:** Simplify fetch example
**Additional Description:**
This PR simplifies the fetch_client example to surface the core APIs,
making it easier to understand how they are supposed to be used.
It also simplifies the building of the engine.

Please let me know your thoughts. I found the original example a bit
intimidating and harder to follow. But I also understand if you want to
illustrate a more advanced use case. I'd be happy to create a separate
`simple_fetch_client` if you'd like to have both.


Risk Level: LOW
It only changes the example.

Testing:
Ran the example on macOS with:
```
./bazelw run //examples/cc/fetch_client -- https://www.google.com/generate_204 https://www.google.com/generate_204
```

Signed-off-by: Vinicius Fortuna <fortuna@users.noreply.github.com>
Signed-off-by: Kuat Yessenov <kuat@google.com>
Co-authored-by: Kuat <kyessenov@users.noreply.github.com>
Co-authored-by: ohadvano <49730675+ohadvano@users.noreply.github.com>