Feature: support SetUpstreamOverrideHost in golang filter (#39754)
Support set upstream override host in golang filter, developer can call
SetUpstreamOverrideHost in DecoderFilterCallbacks, here is the
description for SetUpstreamOverrideHost:

Sets an upstream address override for the request. When the overridden
host exists in the host list of the routed cluster
and can be selected directly, the load balancer bypasses its algorithm
and routes traffic directly to the specified host.

The strict flag determines whether the HTTP request must strictly use
the overridden destination.
When the destination is unavailable/not-found, if strict is set to true,
Envoy responds with a 503 Service Unavailable error;
if strict is set to false, Envoy will fall back to its load balancing
mechanism.

Notice: host must be a valid IP address(with port).

Here is the function:
SetUpstreamOverrideHost(host string, strict bool) error


Signed-off-by: duxin40 <duxin40@gamil.com>