dynamic_modules: adds support for LocalityLbEndpoints metadata (#40021)
Commit Message: dynamic_modules: adds support for LocalityLbEndpoints
metadata
Additional Description:

This adds a new enum variant to
envoy_dynamic_module_type_metadata_source to get LocalityLbEndpoints.
This also fixes the naming style of the enum so that the variants are in
pascal case. This is a followup on #39805.

Risk Level: low (new code path)
Testing: done
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>