xds: add singleton subscription xDS-TP-based config-sources support (#39943)
Followup to #39781, as part of #24773.
This PR adds the framework for singleton subscriptions to resources that
are defined using the xDS-TP config-sources in the bootstrap.
Next is to convert the current xDS singletons (EDS, RDS, etc.) to use
this functionality.

Risk Level: low - not used
Testing: Added unit tests.
Docs Changes: N/A
Release Notes: N/A
Platform Specific Features: N/A

Part of #24773.

---------

Signed-off-by: Adi Suissa-Peleg <adip@google.com>