ext_proc: handle server error response based on the failure_mode_allow config (#39994)
ext_proc: handle server error response based on the failure_mode_allow
config.

This is to fix: https://github.com/envoyproxy/envoy/issues/38553.

If the server returns a spurious response, with failure_mode_allow set
to true, the message will be ignored and external processing is closed,
and the original request is forwarded to upstream as it is. with
failure_mode_allow set to false, local reply is sent to the downstream,
and the HTTP stream is reset.

---------

Signed-off-by: Yanjun Xiang <yanjunxiang@google.com>