aggregated-cluster: replace data-members with getters (#40086)
Commit Message: aggregated-cluster: replace data-members with getters
Additional Description:
This PR removes the `runtime_` and `random_` data-members and adds a
getter to the members that are in the base class (`ClusterImplBase`).
Note that the getters are const, but return a non-const reference, which
is keeping the "spirit" of the previous implementation - a const Cluster
data member was accessed for non-const operations.

Risk Level: low
Testing: N/A
Docs Changes: N/A
Release Notes: N/A
Platform Specific Features: N/A

Signed-off-by: Adi Suissa-Peleg <adip@google.com>