lua filter: ensure active_request_ is verified before deletion Fixes #38017 (#38863)
Lua filter was removing ":status" and then setting a response body,
which was causing envoy to terminate with a SIGSEGV. The issue occurred
because "onEncodeComplete()" attempted to delete "activerequest" via
"deferredDelete()", but "activerequest" had already been deleted
earlier, leading to a segmentation fault. This fix ensures that
"activerequest" is only deleted once, preventing double deletion. This
is done by verifying that "activerequest" is not null before the
existing check of "activerequest->remotecomplete".


Risk Level: Low

Testing:
I added a new integration test in "lua_integration_test.cc" file
specifically for this bug. Without the fix, the test fails; with it
applied, the test passes successfully.

Fixes #38017

---------

Signed-off-by: Leonor Figueira <leonor.figueira@tecnico.ulisboa.pt>