tls: fix incorrectly cached empty peer certificate values
If tls connection data was queried before it was available, and empty
value was being incorrectly cached, preventing later calls from
getting the correct value.

One way this could be triggered was with an access log entry being
emitted before the TLS peer certificate was received during the TLS
handshake, which can be configured with tcp_proxy.

Fixes #40137

Signed-off-by: Greg Greenway <ggreenway@apple.com>