rbac: refactor matchers from shared-ptr to unique-ptr (#40230)
Commit Message: rbac: refactor matchers from shared-ptr to unique-ptr
Additional Description:
This PR replaces the creation interface of the RBAC matchers. Instead of
returning a shared-ptr it now returns a unique-ptr.
There is no reason to keep a shared points, because the matchers are
owned by the filter-config which is already shared.

Risk Level: low
Testing: N/A
Docs Changes: N/A
Release Notes: N/A
Platform Specific Features: N/A

Signed-off-by: Adi Suissa-Peleg <adip@google.com>