sds: Avoid adding empty trusted_ca in CertificateValidationContextSdsApi (#40231)
This PR fixes an issue where the CertificateValidationContextSdsApi adds
an empty trusted_ca to configurations which did not specify one.

This is a bug as of #39751. Now as far as I can tell this line does
nothing (from
[certificate_validation_context_config_impl.cc](https://github.com/envoyproxy/envoy/pull/39751/files#diff-cf088136dc052ddf1762fb3c96c0e8de472f3031f288e7e300558e6e72c8e129))

```
bool allow_empty_trusted_ca = !context.has_trusted_ca();
```

... Because has_trusted_ca will be true even if you didn't specify one.

This causes an issue if you're using a custom certificate validator that
does not depend on the trusted ca field (such as the SPIFFEValidator).
You would need to disable the feature flag
envoy.reloadable_features.reject_empty_trusted_ca_file in a case where
you shouldn't have to.

Signed-off-by: antoniovleonti <leonti@google.com>