matcher: refactor the matcher to avoid unnecessary wrapper and heap allocation (#40249)
In the previous matcher tree implementation. The action is parsed and
stored with format of a ActionFactoryCb. And once the action is hit we
will call the ActionFactoryCb to create Action instance.

However, if we review all these cases, we will find in most cases, the
actual action content is parsed into a shared pointer. Then the
ActionFactoryCb will capture the shared pointer. When the action is hit,
the ActionFactoryCb will create a temporary object in the heap and copy
the shared pointer into the new temporary object. It's make no sense
except complex logic and low efficient memory usage.

Risk Level: low.
Testing: n/a.
Docs Changes: n/a.
Release Notes: n/a.
Platform Specific Features: n/a.

---------

Signed-off-by: wangbaiping(wbpcode) <wbphub@gmail.com>