Make BackwardsTrace::addrMapping a string_view type (#40331)
Commit Message: Make BackwardsTrace::addrMapping a string_view type
Additional Description: This makes it so in the event of it being
triggered in a tsan context without having been initialized, it still
doesn't malloc anything - it is a tsan error to malloc during a signal.
Risk Level: Only crash-related so pretty safe.

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>