maxmind geoip provider: fix bug for no entry found
The maxmind provider previously only checked if the call to the maxmind
C library retuned an error, but failed to consider the case where it
would not find an entry for an IP but return no error. This could lead
to wrong data being decorated for IPs, such as ASN numbers being asigned
to requests which IP did not fact not belong to that network.

The maxmind library's documentation states here that one should always
check if the MMDB_lookup_result_s found_entry member is true first and
otherwise the struct does not contain meaningful data:
https://github.com/maxmind/libmaxminddb/blob/main/doc/libmaxminddb.md#mmdb_lookup_result_s

Signed-off-by: Alex Sundström <asundstrom@spotify.com>