lua: adds filterState() on streamInfo() (#40023)
## Description

This PR adds support for accessing `filter_state` through the stream
info API. This allows Lua scripts to safely access filter state data,
with proper unpacking of **Any** messages into Lua tables. The
implementation uses same building blocks as that of
`dynamicTypedMetadata()` for which we added the protobuf conversion
utils in the earlier PRs.

---

**Commit Message:** lua: adds `filterState()` on `streamInfo()`
**Additional Description:** Added a new ``filterState()`` on
``streamInfo()`` which could be used to access the filter state from
HTTP filters, such as the ExtProc, etc.
**Risk Level:** Low
**Testing:** Unit tests and integration tests added
**Docs Changes:** Added
**Release Notes:** Added

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>