tests: rename TypeUrl to TestTypeUrl (#40342)
Commit Message: tests: rename TypeUrl to TestTypeUrl
Additional Description:
Tests only - rename the type `TypeUrl` to `TestTypeUrl`.
The reason is that we are going to introduce a TypeUrl class in
[resource_name.h](https://github.com/envoyproxy/envoy/blob/main/source/common/config/resource_name.h)
and it conflicts with the tests type.

This was done by running `grep -rlZ "TypeUrl::get" test | xargs -0 sed
-i 's/TypeUrl::get/TestTypeUrl::get/g'`.

Risk Level: low - tests only
Testing: N/A
Docs Changes: N/A
Release Notes: N/A
Platform Specific Features: N/A

---------

Signed-off-by: Adi Suissa-Peleg <adip@google.com>