Avoid an unnecessary string copy in ProxyFilter (#40431)
Commit Message: Avoid an unnecessary string copy in ProxyFilter
Additional Description: This code was changed recently to fix an issue,
but it gained a string-copy that's avoidable for any addresses that are
not numeric IPv6 addresses.
Risk Level: Negligible.
Testing: Existing tests cover it.
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>