oauth2: add Secure attribute for prefixed cookies on signout
When using cookie names prefixed with `__Secure-` or `__Host-`, browsers
require the `Secure` attribute to be present on the `Set-Cookie` header
for any modification, including deletion.

The OAuth2 filter did not add this attribute when generating
the `Set-Cookie` headers to clear the OAuth cookies. This caused the
browser to ignore the deletion request for prefixed cookies, effectively
preventing the user from being signed out correctly.

This patch fixes the behavior to set the `Secure` attribute when either
prefix is found.

Signed-off-by: Tony Allen <txallen@google.com>
Signed-off-by: Ryan Northey <ryan@synca.io>
Signed-off-by: Boteng Yao <boteng@google.com>