tracing: simplify the custom tags to avoid create new custom tag map for every request (#40366)
In the previous implementation, we will create a new custom tags map for
every request to merge the custom tags from hcm and route.

This new implementation clean up the code and avoid that.

Risk Level: low.
Testing: n/a.
Docs Changes: n/a.
Release Notes: n/a.
Platform Specific Features: n/a.

---------

Signed-off-by: wangbaiping(wbpcode) <wbphub@gmail.com>