xds-federation: adding RDS using on xds-TP based configs (#40425)
This PR makes RDS the second xDS-resource-type to support xds-tp based
configs, which enables it to be used in a federated-xDS scenario.
This feature is under a runtime flag
`envoy.reloadable_features.xdstp_based_config_singleton_subscriptions`
which is disabled by default. Once we get to test it end-to-end, and add
a few more xDS-resource types to the support, it could be enabled by
default.
We removed having the `config_source` constraint in an `RDS` config, and
added a validation to check that the resource is an xDS-TP resource.

Risk Level: low - disabled by default
Testing: Added both unit and integration tests.
Docs Changes: not yet
Release Notes: not yet
Platform Specific Features: N/A

---------

Signed-off-by: Adi Suissa-Peleg <adip@google.com>