deps: bump up cel-cpp to v0.11.0 (#39334)
## Description

This PR bumps up the version of `cel-cpp` to v0.11.0.

There are some major changes where we now have `cel::expr::ParsedExpr`
instead of `google::api::expr::v1alpha1::ParsedExpr`. We are doing the
conversion in the code to keep it backward compatible. We should remove
it in a follow-up PR by migrating everything to use
`cel::expr::ParsedExpr` instead.

---

**Commit Message:** deps: bump up `cel-cpp` to v0.11.0
**Additional Description:** Bump up the version of `cel-cpp` to v0.11.0
**Risk Level:** Low
**Testing:** CI
**Docs Changes:** N/A
**Release Notes:** N/A

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>