Update udp_tunneling_integration_test.cc's unknown capsule type (#40494)
When this test was written, capsule type 0x01 was unknown to Envoy and
Google QUICHE, but it had already been assigned to the connect-ip draft,
later RFC 9484. RFC 9297, which defines capsules in general, defines
capsule types of the form 0x29 * N + 0x17, for integer values of N, as
reserved for testing behavior with unknown capsule types. Setting N=0,
we can just use 0x17 instead of 0x01 and per the RFC we should never
collide with a real capsule assignment again.

https://www.rfc-editor.org/rfc/rfc9484.html#section-12.4
https://www.rfc-editor.org/rfc/rfc9297.html#section-5.4-3

Risk Level: low
Testing: test-only change

Signed-off-by: Alejandro R. Sedeño <asedeno@google.com>