Unify scattered header matchers into ContainsHeader (#40528)
Commit Message: Unify scattered header matchers into ContainsHeader
Additional Description: Across envoy there's `Http::HeaderValueOf`,
`HeaderHasValue`, `HeaderHasValueRef`, `SingleHeaderValueIs` and
`HasHeader` all performing the same task. The only well-named one of
these is `HasHeader`, which is the smallest scoped and worst-implemented
one.

`HeaderHasValueRef` sounds like it's a matcher for a header, checking
that it has a ValueRef, when it is in fact a matcher for a reference to
a HeaderMap, checking that it contains a specified key-value pair.
`HeaderValueOf` *almost* makes sense, but Value is kind of meaningless,
and you have to read it like `HeaderValueOf(x, y)` means
`HeaderValueOf(x) is (y)` which is not a very natural reading.
`HeaderHasValue` again sounds like it's a matcher for a header, checking
that it has a value, when it is in fact a matcher for a HeaderMap
pointer, checking that it contains a specified key-value pair.
`SingleHeaderValueIs` is kind of okay in that by specifying single it
sort of implies the matcher *isn't* operating on only one header because
if it was why would you say that, but ugh. Also lived in a limited
namespace, twice.
`HasHeader` is doing what it says on the tin but has relatively
unhelpful output on failure and lived in a limited namespace.
`HasNoHeader` is kind of alright but unnecessary to be its own thing,
and has relatively unhelpful output on failure, and lived in a limited
namespace.

Replacing all of these is `ContainsHeader`, which is intuitively
checking that a HeaderMap *contains* the header described by a key-value
pair. Initially I also replaced `HeaderHasValue` with
`PointeeContainsHeader()` but it turns out there were in fact zero uses
of `HeaderHasValue`, and `Pointee(ContainsHeader())` can achieve the
same thing so it's not necessary.
Risk Level: May require external extensions to update their matchers.
Testing: It's test-only and used in existing tests.
Docs Changes: Yes, HeaderValueOf documentation has been replaced.
Release Notes: Yes.
Platform Specific Features: n/a

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>