Merging implementations of Strict and Logical DNS (#38483)
The main idea here is tackling #37145 (which is a follow up from
#36353).
 
This approach simplifies the code merging the two current implementation
of DNS cluster (logical and strict) by removing unnecessary logical DNS,
as it can be treated as a special case of strict DNS.

According to the documentation:

_When using **strict DNS** service discovery, Envoy will continuously
and asynchronously resolve the specified DNS targets. Each returned IP
address in the DNS result will be considered an explicit host in the
upstream cluster. This means that if the query returns three IP
addresses, Envoy will assume the cluster has three hosts, and all three
should be load balanced to. If a host is removed from the result Envoy
assumes it no longer exists and will drain traffic from any existing
connection pools_

_**Logical DNS** uses a similar asynchronous resolution mechanism to
strict DNS. However, instead of strictly taking the results of the DNS
query and assuming that they comprise the entire upstream cluster, a
logical DNS cluster only uses the first IP address returned when a new
connection needs to be initiated. Thus, a single logical connection pool
may contain physical connections to a variety of different upstream
hosts. Connections are never drained, including on a successful DNS
resolution that returns 0 hosts._

---------

Signed-off-by: Gustavo <grnmeira@gmail.com>