rbac: switch to using LC Trie for RBAC IP range matcher (#40536)
## Description

This PR is a follow-up to our [previous
change](https://github.com/envoyproxy/envoy/pull/40493) where we
introduced the capability for RBAC to use LC-Trie. This change is for
switching the RBAC to using the LC-Trie.

We have also deprecated the old constructor in this PR in favor of
`absl::Status`.

---

**Commit Message:** rbac: switch to using LC Trie for RBAC IP range
matcher
**Additional Description:** This change is for switching the RBAC to
using the LC-Trie
**Risk Level:** Low
**Testing:** Added Unit Tests
**Docs Changes:** N/A
**Release Notes:** N/A

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>