dynamic_modules: add support for user-defined metrics
Similar to how it is done in wasm:
- DynamicModuleHttpFilterConfig now holds a Stats::Scope. All metrics
declared in this scope are placed in a custom stat namespace so as to
not collide with envoy-defined metrics. Using
Stats::CustomStatNamespaces, this namespace prefix is stripped from all
stat names before they are emitted, e.g. to Prometheus format.
- Add new ABI "method"s on
envoy_dynamic_module_http_filter_envoy_config_ptr that allow metrics to
be declared in the scope of the filter config. These methods return
opaque pointers to Stats::{Counter,Gauge,Histogram} that allow said
declared metrics to be manipulated. Right now, only the basics of each
metric type are exposed in the ABI. For example, you can't read metric
values, or reset metrics to 0, or...
- Add corresponding wrappers to the Rust SDK.
- Add tests, both against the C++ ABI implementation, and against the
Rust SDK.

Signed-off-by: Ian Kerins <git@isk.haus>