maxmind geoip provider: fix bug for no entry found (#40365)
The maxmind provider previously only checked if the call to the maxmind
C library returned an error, but failed to consider the case where it
would not find an entry for an IP but return no error. This could lead
to wrong data being decorated for IPs, such as ASN numbers being
assigned to requests which IP did not actually belong to that network.

The maxmind library's documentation states here that one should always
check if the MMDB_lookup_result_s found_entry member is true first and
otherwise the struct does not contain meaningful data:
https://github.com/maxmind/libmaxminddb/blob/main/doc/libmaxminddb.md#mmdb_lookup_result_s

Some of the maxmind provider tests seem to have accidentally relied on
the buggy behaviour. To fix them I had to get proper data from the
database files and in a few cases make adjustments to the test data
itself (like for the tests around updating the database). My methodology
was something like this:
1. Look through the source data for the test data here[
https://github.com/maxmind/MaxMind-DB/tree/main/source-data](https://github.com/maxmind/MaxMind-DB/tree/main/source-data)
for IPs that would fit the test cases
2. For cases where data was expected to exist in 2 different databases,
like the City and ASN ones, I used a small python script that would find
IPs that were in ranges present in both files
3. For the tests around updating the database I had to dump the existing
test data into json using[
https://github.com/maxmind/MaxMind-DB-Reader-perl/blob/main/eg/mmdb-dump-database](https://github.com/maxmind/MaxMind-DB-Reader-perl/blob/main/eg/mmdb-dump-database)
, manually change some values and then write them to the mmdb file
format again with[
https://github.com/maxmind/MaxMind-DB/tree/main/cmd/write-test-data](https://github.com/maxmind/MaxMind-DB/tree/main/cmd/write-test-data)

Disclosure: this was initially submitted to the envoy security group
first but as the geoip filter and maxmind provider are alpha extensions
they aren’t covered by the security process and I was told it was fine
to just submit a fix.

---
Commit Message: maxmind geoip provider: fix bug for no entry found
Additional Description: As per the maxmind library’s docs if the value
of the found_entry field in the lookup result isn’t checked it’s
possible to pull useless data from the DB
https://github.com/maxmind/libmaxminddb/blob/main/doc/libmaxminddb.md#mmdb_lookup_result_s
Risk Level: low
Testing: unit testing
Docs Changes: N/A
Release Notes: Added
Platform Specific Features: N/A

---------

Signed-off-by: Alex Sundström <asundstrom@spotify.com>
Signed-off-by: Alex Sundström <alexnssm@gmail.com>
Co-authored-by: Alex Sundström <asundstrom@spotify.com>